error caused by renaming resolved
